+++ src/commonMain/kotlin/csense/kotlin/extensions/primitives/String.kt
@@ -334,10 +334,18 @@ public inline fun String.isNewLine(): Boolean = when (this.length) {
 }
 
 /**
- * Returns null if this string is empty
- * @return [String] null if this string is empty. Otherwise, the string
+ * Returns null if this string [isEmpty]
+ * @return [String] null if this string [isEmpty]. Otherwise, the string
  */
 public inline fun String.nullOnEmpty(): String? = ifEmpty {
+    null
+}
+
+/**
+ * Returns null if this string [isBlank]
+ * @return [String] null if this string [isBlank]. Otherwise, the string
+ */
+public inline fun String.nullOnBlank(): String? = ifBlank {
     null
 }