+++ src/commonMain/kotlin/csense/kotlin/specificExtensions/boolean/BooleanMappings.kt
@@ -1,0 +1,9 @@
+package csense.kotlin.specificExtensions.boolean
+
+import kotlin.jvm.*
+
+@JvmInline
+public value class BooleanMappings(public val boolean: Boolean)
+
+public inline val Boolean.mappings: BooleanMappings
+    get() = BooleanMappings(this)