+++ src/jvmMain/kotlin/csense/kotlin/specificExtensions/boolean/BooleanMappingsJvm.kt
@@ -1,0 +1,11 @@
+package csense.kotlin.specificExtensions.boolean
+
+import csense.kotlin.extensions.*
+
+
+public fun BooleanMappings.toNewlineOrEmpty(): String {
+    return boolean.map(
+        ifTrue = System.lineSeparator(),
+        ifFalse = ""
+    )
+}